# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: TeamFlow
  version: '1.0'
servers:
  -
    url: 'http://teamflow-be.gplans.it'
    description: 'API Server'
paths:
...
  /api/account/all:
    get:
      tags:
        - account
      summary: 'Get all accounts data'
      operationId: getAllAccounts
      parameters:
        -
          name: limit
          in: query
          description: 'Limit of elements'
          required: false
          schema:
            type: integer
            example: 10
        -
          name: page
          in: query
          description: 'Current page'
          required: false
          schema:
            type: integer
            example: 1
        -
          name: sortBy
          in: query
          description: 'sort by element'
          required: false
          schema:
            type: string
            example: username
        -
          name: sortValue
          in: query
          description: 'sorting type'
          required: false
          schema:
            type: string
            example: ascend
        -
          name: filterBy
          in: query
          description: 'fields to filter'
          required: false
          schema:
            type: string
            example: name
        -
          name: filterValue
          in: query
          description: 'Values to filter'
          required: false
          schema:
            type: string
            example: ma
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AccountsResponse'
                  message:
                    type: string
                type: object
        401:
          description: 'Not Authorized'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        403:
          description: Forbitten
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
      security:
        -
          bearerAuth: []
  /api/accounts/username/all:
    get:
      tags:
        - account
      summary: 'Get all accounts usernames'
      operationId: getAllUsernames
      parameters:
        -
          name: limit
          in: query
          description: 'Limit of elements'
          required: false
          schema:
            type: integer
            example: 10
        -
          name: page
          in: query
          description: 'Current page'
          required: false
          schema:
            type: integer
            example: 1
        -
          name: sortBy
          in: query
          description: 'sort by element'
          required: false
          schema:
            type: string
            example: username
        -
          name: sortValue
          in: query
          description: 'sorting type'
          required: false
          schema:
            type: string
            example: ascend
        -
          name: filterBy
          in: query
          description: 'fields to filter'
          required: false
          schema:
            type: string
            example: name
        -
          name: filterValue
          in: query
          description: 'Values to filter'
          required: false
          schema:
            type: string
            example: ma
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AccountsUsernames'
                  message:
                    type: string
                type: object
        401:
          description: 'Not Authorized'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        403:
          description: Forbitten
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
      security:
        -
          bearerAuth: []
  '/api/account/{accountId}':
    get:
      tags:
        - account
      summary: 'Get data of a single account'
      operationId: getAccount
      parameters:
        -
          name: accountId
          in: path
          description: 'model id'
          required: true
          schema:
            type: string
            example: 59e7626e-49bf-4e63-b195-f4d826af8ad8
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/AccountResponse'
                  message:
                    type: string
                type: object
        401:
          description: 'Not Authorized'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        403:
          description: Forbitten
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        404:
          description: 'Not Found'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Account not found'
                type: object
      security:
        -
          bearerAuth: []
    put:
      tags:
        - account
      summary: 'Update a specific account'
      description: ''
      operationId: updateAccount
      parameters:
        -
          name: accountId
          in: path
          description: 'model id'
          required: true
          schema:
            type: string
            example: c7817f8f-3539-46c9-a5ec-b85227bb4e07
      requestBody:
        description: 'JSON with account data'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountBodyReq'
      responses:
        200:
          description: Updated
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items: []
                  message:
                    type: string
                    example: 'Update Success'
                type: object
        400:
          description: 'Bad request'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Bad request'
                type: object
        401:
          description: 'Not Authorized'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        403:
          description: Forbitten
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        404:
          description: 'Inserted model not exists'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Account not found'
                type: object
        500:
          description: 'Server Error'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Insert unsuccess'
                type: object
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - account
      summary: 'Delete an existing account'
      description: ''
      operationId: deleteAccount
      parameters:
        -
          name: accountId
          in: path
          description: 'Account id'
          required: true
          schema:
            type: string
            example: c7817f8f-3539-46c9-a5ec-b85227bb4e07
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Delete success'
                type: object
        400:
          description: 'Bad request'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Bad request'
                type: object
        401:
          description: 'Not Authorized'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        403:
          description: Forbitten
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        404:
          description: 'Not found'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Account not found'
                type: object
        500:
          description: 'Server Error'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Delete unsuccess'
                type: object
      security:
        -
          bearerAuth: []
  '/api/account/{accountId}/info':
    get:
      tags:
        - account
      summary: 'Get info of a single account'
      operationId: getAccountInfo
      parameters:
        -
          name: accountId
          in: path
          description: 'model id'
          required: true
          schema:
            type: string
            example: 59e7626e-49bf-4e63-b195-f4d826af8ad8
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/AccountInfoResponse'
                  message:
                    type: string
                type: object
        401:
          description: 'Not Authorized'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        403:
          description: Forbitten
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        404:
          description: 'Not Found'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Account not found'
                type: object
      security:
        -
          bearerAuth: []
  /api/account:
    post:
      tags:
        - account
      summary: 'Create new account'
      description: ''
      operationId: createAccount
      requestBody:
        description: 'JSON with account data'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountBodyReq'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                properties:
                  data:
                    properties:
                      id:
                        type: string
                        example: 1a2d4d27-43d1-4450-9ed0-5c4e7b802dae
                    type: object
                  message:
                    type: string
                    example: 'Account created successfully'
                type: object
        400:
          description: 'Bad request'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Bad request'
                type: object
        401:
          description: 'Not Authorized'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        403:
          description: Forbitten
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        500:
          description: 'Server Error'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Insert unsuccess'
                type: object
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    AccountResponse:
      properties:
        id:
          type: string
          example: 59e7626e-49bf-4e63-b195-f4d826af8ad8
        username:
          type: string
          example: agent
        name:
          type: string
          example: Luca
        surname:
          type: string
          example: 'La Porta'
        email:
          type: string
          example: agent@dev.com
      type: object
    AccountInfoResponse:
      properties:
        name:
          type: string
          example: Luca
        surname:
          type: string
          example: 'La Porta'
      type: object
    AccountsUsernames:
      properties:
        id:
          type: string
          example: a6ff9885-6166-4375-979a-bad6e2b4b9be
        username:
          type: string
          example: luca10
      type: object
    AccountsResponse:
      properties:
        current_page:
          type: number
          example: 1
        data:
          type: array
          items:
            $ref: '#/components/schemas/AccountResponse'
        first_page_url:
          type: string
          example: 'http://localhost:8000/api/account/all?page=1'
        from:
          type: number
          example: 1
        last_page:
          type: number
          example: 1
        last_page_url:
          type: string
          example: 'http://localhost:8000/api/account/all?page=1'
        links:
          type: array
          items:
            properties:
              url:
                type: string
                format: nullable
              label:
                type: string
                example: pagination.previous
              active:
                type: boolean
            type: object
        next_page_url:
          type: string
          format: nullable
        path:
          type: string
          example: 'http://localhost:8000/api/account/all'
        per_page:
          type: number
          example: 10
        prev_page_url:
          type: string
          format: nullable
        to:
          type: number
          example: 1
        total:
          type: number
          example: 1
      type: object
    AccountBodyReq:
      properties:
        username:
          type: string
          example: john10
        name:
          type: string
          example: John
        surname:
          type: string
          example: Doe
        email:
          type: string
          example: johndow@email.com
        password:
          type: string
          example: abc1234
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...